# src/main/java/com/zl/way/h5app/api/WayArticlePostApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen Do not edit the class manually.
 */
package com.zl.way.h5app.api;

import javax.validation.Valid;
import javax.validation.constraints.NotNull;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import com.zl.way.h5app.api.model.WayQueryArticlePost;
import com.zl.way.h5app.api.model.WayQueryArticlePostResp;

@RequestMapping(value = "")
public interface WayArticlePostApi {

    @RequestMapping(value = "/app/article/{postId}", produces = {"application/json"}, method = RequestMethod.GET)
    ResponseEntity<WayQueryArticlePost> getArticle(@PathVariable("postId") Long postId,
        @RequestHeader(value = "token", required = true) String token,
        @RequestHeader(value = "userLoginId", required = false) Long userLoginId);

    @RequestMapping(value = "/app/article", produces = {"application/json"}, method = RequestMethod.GET)
    ResponseEntity<WayQueryArticlePostResp> queryArticles(
        @NotNull @Valid @RequestParam(value = "page", required = true, defaultValue = "1") Integer page,
        @NotNull @Valid @RequestParam(value = "pageSize", required = true, defaultValue = "10") Integer pageSize,
        @Valid @RequestParam(value = "keywords", required = false) String keywords);

}